Assignment 3/Bank Agent/pyproject.toml
[project]
name = "bank-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.7",
    "pydantic>=2.11.7",
]
